# src/frontend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "VenueAxe.Web | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5280/"
    }
  ],
  "paths": {
...
    "/api/admin/waivers/search": {
      "get": {
        "tags": [
          "Waivers"
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "term",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaiverDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaiverDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaiverDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/waivers/{id}/pdf": {
      "get": {
        "tags": [
          "Waivers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/waivers/templates/venue/{venueId}": {
      "get": {
        "tags": [
          "Waivers"
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaiverTemplateDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaiverTemplateDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaiverTemplateDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/waivers/templates/{templateId}": {
      "put": {
        "tags": [
          "Waivers"
        ],
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWaiverTemplateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWaiverTemplateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWaiverTemplateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WaiverTemplateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaiverTemplateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaiverTemplateDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateWaiverTemplateRequest": {
        "required": [
          "title",
          "bodyTextMarkdown"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "bodyTextMarkdown": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "WaiverDto": {
        "required": [
          "id",
          "venueId",
          "bookingId",
          "signerFirstName",
          "signerLastName",
          "signerEmail",
          "signerPhone",
          "dateOfBirth",
          "isGuardianSigning",
          "minorsCoveredJson",
          "signatureImagePngBase64",
          "signedAtUtc",
          "expiresAtUtc",
          "isExpired"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "bookingId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "signerFirstName": {
            "type": "string"
          },
          "signerLastName": {
            "type": "string"
          },
          "signerEmail": {
            "type": "string"
          },
          "signerPhone": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "isGuardianSigning": {
            "type": "boolean"
          },
          "minorsCoveredJson": {
            "type": [
              "null",
              "string"
            ]
          },
          "signatureImagePngBase64": {
            "type": "string"
          },
          "signedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "isExpired": {
            "type": "boolean"
          }
        }
      },
      "WaiverTemplateDto": {
        "required": [
          "id",
          "venueId",
          "versionNumber",
          "title",
          "bodyTextMarkdown",
          "sha256Hash"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "versionNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "bodyTextMarkdown": {
            "type": "string"
          },
          "sha256Hash": {
            "type": "string"
          }
        }
      }
    }
  },
...
}